+++ test/codeception/tests/acceptance/DefaultWPCest.php
@@ -13,7 +13,7 @@ class DefaultWPCest
     }
 
     // tests
-    public function viewPageHHVMNoCookie(AcceptanceTester $I)
+    public function viewPageHHVM(AcceptanceTester $I)
     {
         $I->amOnUrl('http://hhvm.hgv.test');
         $I->amOnPage('/');
@@ -45,158 +45,4 @@ class DefaultWPCest
         $I->seeCurrentUrlEquals('/');
         $I->seeResponseCodeIs(200);
     }
-
-    // Test that can load the default WP with PHP 5.5
-    public function viewPageCookiePHP(AcceptanceTester $I)
-    {
-        $I->setCookie('backend', 'php');
-        $I->amOnUrl('http://hhvm.hgv.test');
-        $I->amOnPage('/');
-        $I->seeCurrentUrlEquals('/');
-        $I->seeResponseCodeIs(200);
-        $I->assertRegExp('#PHP/5\.5\.(.*)#', $I->grabHttpHeader('X-Powered-By'));
-    }
-
-    // Test that can load the default WP with PHP 5.5
-    public function viewPageCookiePHP5(AcceptanceTester $I)
-    {
-        $I->setCookie('backend', 'php5');
-        $I->amOnUrl('http://hhvm.hgv.test');
-        $I->amOnPage('/');
-        $I->seeCurrentUrlEquals('/');
-        $I->seeResponseCodeIs(200);
-        $I->assertRegExp('#PHP/5\.5\.(.*)#', $I->grabHttpHeader('X-Powered-By'));
-    }
-
-    // Test that can load the default WP with PHP 5.5
-    public function viewPageCookiePHP55(AcceptanceTester $I)
-    {
-        $I->setCookie('backend', 'php55');
-        $I->amOnUrl('http://hhvm.hgv.test');
-        $I->amOnPage('/');
-        $I->seeCurrentUrlEquals('/');
-        $I->seeResponseCodeIs(200);
-        $I->assertRegExp('#PHP/5\.5\.(.*)#', $I->grabHttpHeader('X-Powered-By'));
-    }
-
-    // Test that can load the default WP with PHP 5.6
-    public function viewPageCookiePHP56(AcceptanceTester $I)
-    {
-        $I->setCookie('backend', 'php56');
-        $I->amOnUrl('http://hhvm.hgv.test');
-        $I->amOnPage('/');
-        $I->seeCurrentUrlEquals('/');
-        $I->seeResponseCodeIs(200);
-        $I->assertRegExp('#PHP/5\.6\.(.*)#', $I->grabHttpHeader('X-Powered-By'));
-    }
-
-    // Test that can load the default WP with PHP 7
-    public function viewPageCookiePHP7(AcceptanceTester $I)
-    {
-        $I->setCookie('backend', 'php7');
-        $I->amOnUrl('http://hhvm.hgv.test');
-        $I->amOnPage('/');
-        $I->seeCurrentUrlEquals('/');
-        $I->seeResponseCodeIs(200);
-        $I->assertRegExp('#PHP/7\.(.*)#', $I->grabHttpHeader('X-Powered-By'));
-    }
-
-    // Test that can load the default WP with HHVM
-    public function viewPageCookieHHVM(AcceptanceTester $I)
-    {
-        $I->setCookie('backend', 'hhvm');
-        $I->amOnUrl('http://hhvm.hgv.test');
-        $I->amOnPage('/');
-        $I->seeCurrentUrlEquals('/');
-        $I->seeResponseCodeIs(200);
-        $I->assertRegExp('#HHVM(.*)#', $I->grabHttpHeader('X-Powered-By'));
-    }
-
-    /**
-     * This test will check the file upload size limit in wp-admin media uploader
-     */
-    public function checkFileUploadLimitHHVM(AcceptanceTester $I)
-    {
-        $I->wantTo('Log into site and check file upload on media page');
-        $I->amOnUrl('http://hhvm.hgv.test');
-        $I->amOnPage('/wp-login.php');
-        $I->fillField('log', 'wordpress');
-        $I->fillField('pwd', 'wordpress');
-        $I->click('Log In');
-
-        $I->amOnPage('/wp-admin/media-new.php');
-        $I->see('Maximum upload file size: 200 MB.');
-    }
-
-    /**
-     * This test will check the file upload size limit in wp-admin media uploader
-     */
-    public function checkFileUploadLimitPHP(AcceptanceTester $I)
-    {
-        $I->wantTo('Log into site and check file upload on media page');
-        $I->amOnUrl('http://php.hgv.test');
-        $I->amOnPage('/wp-login.php');
-        $I->fillField('log', 'wordpress');
-        $I->fillField('pwd', 'wordpress');
-        $I->click('Log In');
-
-        $I->amOnPage('/wp-admin/media-new.php');
-        $I->see('Maximum upload file size: 200 MB.');
-    }
-
-    /**
-     * This test will check the file upload size limit in wp-admin media uploader
-     */
-    public function checkFileUploadLimitPHP7(AcceptanceTester $I)
-    {
-        $I->wantTo('Log into site and check file upload on media page');
-        $I->setCookie('backend', 'php7');
-        $I->amOnUrl('http://php.hgv.test');
-        $I->amOnPage('/wp-login.php');
-        $I->assertRegExp('#PHP/7\.(.*)#', $I->grabHttpHeader('X-Powered-By'));
-        $I->fillField('log', 'wordpress');
-        $I->fillField('pwd', 'wordpress');
-        $I->click('Log In');
-
-        $I->amOnPage('/wp-admin/media-new.php');
-        $I->see('Maximum upload file size: 200 MB.');
-    }
-
-    /**
-     * Cause a deliberate 500 error and verify the page does not redirect because not using the cookie.
-     */
-    public function errorPage500no(AcceptanceTester $I)
-    {
-        $file = '/nas/wp/www/sites/hhvm/test-error-500.php';
-        $content = '<?php header("HTTP/1.1 500 Internal Server Error");';
-        $I->writeToFile($file, $content);
-
-        $I->wantTo('Load special error page and get error');
-        $I->amOnUrl('http://hhvm.hgv.test');
-        $I->amOnPage('/test-error-500.php');
-        $I->seeResponseCodeIs(500);
-
-        $I->deleteFile($file);
-    }
-
-    /**
-     * Cause a deliberate 500 error and verify the page is redirected to the special error page.
-     */
-    public function errorPage500yes(AcceptanceTester $I)
-    {
-        $file = '/nas/wp/www/sites/hhvm/test-error-500.php';
-        $content = '<?php header("HTTP/1.1 500 Internal Server Error");';
-        $I->writeToFile($file, $content);
-
-        $I->wantTo('Load special error page and get redirect to our error page');
-        $I->setCookie('backend', 'php7');
-        $I->amOnUrl('http://hhvm.hgv.test');
-        $I->amOnPage('/test-error-500.php');
-        $I->seeResponseCodeIs(500);
-        $I->seeCurrentUrlEquals('/test-error-500.php');
-        $I->see('You are seeing this because you are using the PHP selector, and the page you are working on just returned an error.');
-
-        $I->deleteFile($file);
-    }
-
 }
